Add support for comments in build spec files
Some kind of comment support is important for adding descriptive notes
and documentation inline with build instructions.

Treat lines starting with "# " as a comment.  This is a little quirky in
that the "#" must not appear after other things and must be followed by
a space, but it easy to implement that way and will do well enough for
now.