Improve error message for missing cc-run command
When the command for cc-run is missing, the usage block isn't shown and
instead the "env" command prints the current environment.  The command
exits successfully.  This failure mode is hard to understand at a glance
if it is noticed at all.

When the command is missing in such a way that "env" would print its
environment, print an error and exit with non-zero status.