Mount overlayfs instead of copying directories
Copying directories is very slow for the image sizes we're seeing (354MB
for Xenial).

Instead of creating containers from images with "cp -a", mount a
writable layer over the image files with overlayfs.  This speeds time to
create a Xenial container from 4.7 sec cold (2.2 sec warm) to 0.03 sec
(0.017 sec warm).

Switch to the newer "overlay", available on Linux kernels 4.0 and newer
(available in Xenial/16.04 or newer for the host OS) to support multiple
lowerdir layers, which allows this mounting scheme to support custom
builds.