Expand test suite
The test suite consists of running an example script inside Vagrant.
This has gotten us as far as here, but makes it hard to automate
exercise of edge cases.  Also, in an effort to be realistic, it runs
some fairly slow use-cases.

Create a basic framework for adding additional test cases.  Report on
test pass/fail counts as well as checks within the test case.  Provide a
library of testing functions.  Avoid spewing lots of debug info from
commands so that the test output is tidy.  Rename tests/run to
tests/run-in-vagrant, so that tests/run can be run if already in Vagrant
or in a development environment that directly supports this project.