Allow individual test cases to be run
The cases/example.sh script is a bit slow (~30 sec) because it builds a
real image.  I'd like to run it sometimes as part of a regression suite,
but when focussing on something else, I don't want it to slow me down.

Allow running an individual test file.  Because tests run in the ./tests
subdirectory, but file completion will often be relative to the project
root, strip a "tests/" prefix from the name.