Fix edge case with starting a non-existent image
When starting with a non-existent image, I get:

  mount: special device ignored-by-overlayfs does not exist

And the broken container is left in place such that I can't retry.

This happens because the non-existent image directory is passed right
through to the "mount" command (see lowerdir):

    mount -t overlay \
      -o lowerdir=/var/cookie-cutter/images/non-existent-image/diff \
      -o upperdir=/var/cookie-cutter/containers/tmp.19943/diff \
      -o workdir=/var/cookie-cutter/containers/tmp.19943/work
      ignored-by-overlayfs \
      /var/cookie-cutter/containers/tmp.19943/filesystem

Handle the error by cleaning up the parially created container if one
step fails.