Ensure the "example.sh" case runs first to generate xenial
The xenial base Ubuntu image is expensive to create.  It is created by
the example-script.sh, which is tested by the example.sh case.  We don't
always run the step to create xenial or reset its state when testing,
because it would be too slow.  As a result, the new build.sh case has
been depending on xenial even though it runs before the example.sh case.

Re-order test cases so that xenial will be created before build.sh runs.