Rephrase mount error as missing image error
When starting with a non-existent image, I still see:

  mount: special device ignored-by-overlayfs does not exist

The half-build container is cleaned up, but this is still an obtuse
error.

This happens because the non-existent image directory is passed right
through to the "mount" command (see lowerdir):

    mount -t overlay \
      -o lowerdir=/var/cookie-cutter/images/non-existent-image/diff \
      -o upperdir=/var/cookie-cutter/containers/tmp.19943/diff \
      -o workdir=/var/cookie-cutter/containers/tmp.19943/work
      ignored-by-overlayfs \
      /var/cookie-cutter/containers/tmp.19943/filesystem

Handle the error better by printing a more specific error message that
points to the earlier problem of trying to use an image that doesn't
exist.