Start a debug shell when a build command fails
It's often useful to have the image state available to debug a problem
when a build fails.

One approach would be to not clean up after an error, but this leaves
bad containers around for an unknown amount of time to create clutter.

Instead, start a shell to investigate the problem, and clean up the
container when the shell exits.  Read commands from fd 3 instead of
stdin to leave stdin available for the shell.  Only enable this debug
mode when the "--debug" flag is present.