Add cc-boot to bootstrap a container with init
For many applications, running a single application per container works
nicely.  But for suites of applications that expect to work together in
the same virtual machine (VM), it is useful to have an init process that
manages all the applications (such as SMTP server, mail delivery agent,
and IMAP server).

Add a command called cc-boot that parallels cc-run but bootstraps an
init process in a container instead of running a specified command.  Use
systemd-nspawn to start the container, which has the bonus of providing
more isolation that chroot.  Briefly mention systemd tools, such as
machinectl and systemd-cgtop, for manaching such containers.