Add --volume option to cc-boot
In order to make disposable containers useful for stateful applications,
a way to carry state from one container to a future one is needed.

Add support for volumes that can be bind-mounted into a container to
provide access to persistent files whose lifetime extends beyond that of
the container itself.