Use "rm -rf" more carefully by adding "--one-file-system"
I lost some data outside of a container when unmounting a volume failed
and the container was cleaned up with rm -rf.

Use --one-file-system so that file removal will not recurse into a mount
point.  This seems to protect both bind-mounts and overlay mounts,
though the former is more the concern.  Also, in theory this is not
necessary for removal of things other than containers (because they are
what have bind-mounts in them), but to be safe, just use it everywhere
that "-rf" is used... maybe someday someone will copy-paste or add an
argument to one of those that _is_ a container.