Add support for mounting individual files as volumes
Linux refuses to mount a volume atop a non-existent path in
the container.  We create the directory if it doesn't exist.
However, if we are mounting an individual file, that auto-created
directory causes a file type mismatch:

    mount: /path/to/file: mount(2) system call failed: Not a directory.

Auto-create an empty file if we intend to mount a file, while still
auto-creating an empty dir if we intend to mount a dir.  Raise an
explicit error if the file we intend to mount is neither.