Configure Ubuntu images for security patching by default
Currently, Ubuntu images produced by cc-build-ubuntu-image do not have
security patching enabled.  They are produced by debootstrap, which has
a constraint of only allowing one package source by default -- no security
updates.

Add a security package source and run "apt-get dist-upgrade" to apply
all available updates.  While at it, add package sources for the
components "universe", "restricted", and "multiverse".  And add package
sources for distro updates which aren't required for security but may
improve stability of the system.  This brings the produced image closer
to inline with what is recommended by Ubuntu (to include security and
updates) and available by default in Vagrant, DigitalOcean, and Docker
images for Ubuntu (which still differs in that it includes backports).
Finally, extract separate helper scripts for the original debootstrap
process and for the addition of security patches and components.