Refine output of waiting for process to stop
When cc-run's cleanup code runs on exit, it fails to output a newline
at the end, which makes it a bit hard to read on the terminal.

Add a newline at the end, and also add a space before the waiting dots
to clarify that, when only one is printed, it's not just a period.