Add test coverage for lingering child process in cc-run
Sometimes a process under cc-run will fork a child (or children) that
require time to shutdown gracefully, even after the parent exits.

Add test coverage to show that cc-run will try to wait for the children
to unmount, so that unmounting doesn't fail.  Also refine the
"slow-existing" test case's output.