Unmount after mounting /proc in cc-run
Sometimes it's useful to mount /proc from within a container.  For
example, Metabase (and possibly any java app) seems to require /proc.
(Although this definitely opens security issues, security was not great
with chroot running as root anyway, and a lot of the value from
cookie-cutter is in isolating dependencies and creating environments
repeatably.)  Unfortunately, cc-run is not cleaning up these filesystems
mounted from within.

Although we could grep /proc/mounts for submounts, umount already
supports a "--recursive" option that does what we need and also obviates
a bunch of volume cleanup code, so use that instead.  It seems to be
smart enough not to unmount (for example) a bunch of previouly-mounted
directories if I mount "/" as a volume (as a backup container might).