Accept name and "--rm" independently
I'd like to be able to pass a container name to cc-run and still have
the container cleaned up sometimes.

Instead of acceting "--rm" in lieu of a name, accept it whether a name
is present or not, and accept a name whether "--rm" is present or not.

This is a breaking change.  To upgrade: add "--name" before all
container names in calls to cc-{run,boot} that don't pass "--rm".