Leave pre-existing container when cc-run has a name collision
Currently, if cc-run is repeatedly called with a name and no "--rm", it
will fail every other time.  The first time will create the container,
and the 2nd time will try to create the container, fail because it
exists, and then clean up after itself by removing the container by that
name (which is from a prior run).  The process repeats.  This should not
happen: cc-run should only clean up after state from the current run.

When cc-run is called with the name of a container that already exists:

* Don't automatically remove the container as part of the failure clean
  up code.

* Only out put one line of error text, not four "mkdir" failure messages