Assume '--rm' if no name is passed to cc-run
When calling cc-run without a name, it's hard to find the container that
was created, so we should remove it by default.  Otherwise these
containers will accumulate over time and fill the disk.

Assume '--rm' when no '--name' is given to cc-{run,boot}.  We can always
add a '--no-rm' later if needed.