Use namespaces for isolation in addition to chroot
With cc-run, we currently only use chroot for isolation.  This is great
for isolating APT package dependencies and file systems, but it's not
great for security, especially if the process(es) in the container run
as root.  Also, using mount and pid namespaces might make some kind of
container cleanup more reliable.

Use "unshare" (in cc-run) and "nsenter" (in cc-exec) to interact with
the namespaces.  Track the pid of cc-run process so that we can
reassociate with its pid namespace in cc-exec.  The persisting and
re-entering of the pid namespace via file doesn't seem to work.