Build images and containers from immutable images
If an image is rebuilt while a continer is using it, the files within
the container will be affected.  Since the goal of containers is to
isolate the inside from the outside, this is not a desirable or
controlled way of introducing updates to underlying images.  A similar
problem exists for images: if image B builds on image A, and image A is
updated, image B will be changed in possibly unpredictable ways.

Instead, allow underlying image versions to continue to exist immutably
after the image is rebuilt.  Reference these versions by a logial name,
implemented as a mutable symlink that points to the immutable image
version.  When starting a new container, instead of recording the
logical name, record the immutable image version.