Nest image versions under image dirs
The proliferation of image versions is making the images/ directory
unwieldy, makes it hard to reliably distinguish an image version from a
regular image, introduces the task of extracting the name from the
version, and makes it hard to see the disk usage of all versions of a
particular image using unix tools like "du".

Nest image versions within their image directory.

Because the tests do not recreate the xenial image every time because it
is too slow, this command must be run to make them pass after adding or
removing this commit.  (However, it isn't necessary to regenerate all
images in the new format for the "cc-" commands to work.)

    cc-build-ubuntu-image xenial