Avoid unsharing user namespace in cc-run
Unsharing the user namespace is causing some problems because root
inside the container is not really root:

- Call "runuser -u nobody" in the container
- Read or write to root-only files on the filesystem
- Install packages (due to the above two issues)

Although there is appeal to having root in the container be truly
unprivileged outside, I'm not sure how to address these issues currently
with "--map-root-user".

Omit "--map-root-user" and "--user" for now.  Some features in 20.04 and
forthcoming in 22.04's unshare versions might possibly help; we can
revisit then.  In the mean time, we can continue living with "runuser"
to drop privs in the container.