Form anonymous container names from image not "tmp"
We are currently naming anonymous containers (without "--name") with a
prefix of "tmp" and the pid.  This is not quite right now that the
"--rm" flag is independent of "--name".

Instead, name it after the image that the container was created from.

I tried also using a random string suffix instead of the shell pid, but
it adds a significant amount of runtime, so stick with pid for now:

    $ time tr -dc 'a-zA-Z0-9' </dev/urandom | head -c 8
    real    0m0.013s
    user    0m0.000s
    sys     0m0.012s

Continue using "." instead of "-", as it seems less likely to cause
naming collisions with a named container.