Add garbage collection of old image versions
If we keep on creating new image versions, eventually old unused
versions will fill the disk.

Automatically reclaim unused image versions, except for the current
current version (which must be kept even if not yet used).  Although we
_could_ walk all containers and derived images to determine what to
collect, maintain an index of references so that the GC process is fast
enough to be done whenever a new image version is saved by cc-save.
Create entries in the index of references whenever creating a new
container or derived image using hard-links, so that it is easy to check
freshness of the references and they are robust to "rm -rf"-ing a
container or old image.