Reclaim a named container directory if its process finished
When running named containers as system services, if the system crashes
ungracefully, leaving the container directories around, then those
services will fail to start on boot.  I'd like a reboot to allow
these kinds of services to easily resume.

Instead of failing to start a container when a name is taken, first
check if the process has stopped (as evidenced by a stale pid file).  If
yes, then reclaim that directory: remove it for reuse.  If the process
is still running, then it is still an error to start another container
by the same name.  If the pid file is absent, then perhaps two process
are trying to create the same container concurrently: this is also still
an error.  Or, it is remotely possible that things got fouled up some
other way, for example, container removal failed, and, while it might be
ideal to recover from this situation, we currently also treat it as an
error and refuse to start the container.

Also, extract cc-stop and cc-remove to be used multiple places.