Preserve pid file so container remains eligible for reclaim
Postgres sometimes fails to stop cleanly with errors like this:

    root@postoffice-18-04:/usr/local/oak-cookies# systemctl stop postgres-container.service ; systemctl status postgres-container.service | cat
    ● postgres-container.service - Postgres database running in a container
       Loaded: loaded (/etc/systemd/system/postgres-container.service; enabled; vendor preset: enabled)
       Active: inactive (dead) since Tue 2022-04-05 02:03:55 UTC; 12ms ago
      Process: 16538 ExecStart=/bin/bash -c     set -x;    cc-run      -v /mnt/postgres/postgres-data:/var/lib/postgresql/      --rm --name postgres      postgres      runuser -u postgres -- sh -c '        trap "echo sigchild" CHLD &&        stat_tmp=/var/run/postgresql/10-main.pg_stat_tmp/ &&        [ -d "$stat_tmp" ] || mkdir "$stat_tmp" &&        /usr/lib/postgresql/10/bin/postgres          -D /var/lib/postgresql/10/main          -c config_file=/etc/postgresql/10/main/postgresql.conf      '      2>&1   (code=killed, signal=TERM)
     Main PID: 16538 (code=killed, signal=TERM)

    Apr 05 02:03:55 postoffice-18-04 postres-container[16538]: Could not kill process 16624: No such process
    Apr 05 02:03:55 postoffice-18-04 postres-container[16538]: Could not kill process 16627: No such process
    Apr 05 02:03:55 postoffice-18-04 postres-container[16538]: Could not kill process 16628: No such process
    Apr 05 02:03:55 postoffice-18-04 postres-container[16538]: cc-stop: Sent SIGTERM.  Waiting for processes to stop .
    Apr 05 02:03:55 postoffice-18-04 postres-container[16538]: rm: cannot remove '/var/cookie-cutter/containers/postgres/namespaces/pid': Device or resource busy
    Apr 05 02:03:55 postoffice-18-04 postres-container[16538]: rm: cannot remove '/var/cookie-cutter/containers/postgres/namespaces/mount': Device or resource busy
    Apr 05 02:03:55 postoffice-18-04 postres-container[16538]: rm: cannot remove '/var/cookie-cutter/containers/postgres/namespaces/ipc': Device or resource busy
    Apr 05 02:03:55 postoffice-18-04 postres-container[16538]: rm: cannot remove '/var/cookie-cutter/containers/postgres/namespaces/uts': Device or resource busy
    Apr 05 02:03:55 postoffice-18-04 postres-container[16538]: rm: cannot remove '/var/cookie-cutter/containers/postgres/filesystem': Device or resource busy
    Apr 05 02:03:55 postoffice-18-04 systemd[1]: Stopped Postgres database running in a container.

cc-stop is trying to send a SIGTERM, but it's not working.  I
suspect the "Could not kill" messages are because postgres already
exited due to systemd's SIGTERM between when cc-stop found them and
tried to signal them.  I think the "Device or resource busy" lines might
be because systemd is not letting cc-run's cleanup do its thing, which
takes longer because postgres has more work to do at shutdown.

It's not because of the cc-exec metrics processes, because I can stop
those services and the problem still happens with shutting down
postgres.  (I had thought it might be, because they are in the container
but not in systemd's control group.)

On next start, we get an error that the container already exists and
cannot be started.  Normally, as a named container, this would be
eligible for being reclaimed by cc-run, but its $container/pid file has
been removed, which disables the feature (because we cannot see if the
pid is still running, and for legacy containers).

We could configure systemd to call cc-stop to stop the process instead
of sending SIGTERM as usual.  I suspect this would need to be done for
any container that is slow to exit and runs under systemd.  But I don't
want to have to make that judgement call about whether it will be slow
to exit for each service or risk it failing to start.  And any service
could become slow to exit on a busy server.

Instead, change cc-remove to preserve the $container/pid file that makes
a named container eligible for reclaiming at next cc-run until other
files have been removed first.