Clean up after cc-build background processes
When cc-build builds a ruby app, I see these errors:

    umount: /var/cookie-cutter/containers/build.31936/filesystem: target is busy.
    umount: /var/cookie-cutter/containers/build.31936/filesystem: target is busy.

They are because a gpg-agent is still running in the background in the
build container:

    ~# fuser -v -mM /var/cookie-cutter/containers/build.31936/filesystem
                         USER        PID ACCESS COMMAND
    /var/cookie-cutter/containers/build.31936/filesystem:
                         root     kernel mount /var/cookie-cutter/containers/build.31936/filesystem
                         root       3258 Frce. gpg-agent

cc-build runs build commands in a lightweight container for speed, so it
doesn't have the robust cleanup of linux pid namespaces.  Add a call to
some recently developed cleanup steps in cc-stop.